+++ src/components/HeatmapBox/HeatmapBox.tsx
@@ -1,11 +1,22 @@
+import React from "react";
 import { ReactNode } from "react";
-import { Box } from "src/types";
 
 interface HeatmapBoxProps {
-  box: Box;
+  columnIndex: number;
+  rowIndex: number;
+  style: any;
+  data: any;
 }
 
-export function HeatmapBox({ box }: HeatmapBoxProps) {
+const GUTTER_SIZE = 2;
+
+function HeatmapBox({ columnIndex, rowIndex, style, data }: HeatmapBoxProps) {
+  const index = columnIndex * 7 + rowIndex;
+  const box = data[index] ?? {
+    backgroundColor: "transparent",
+    isSpaceBetweenBox: true,
+  };
+
   const boxClassNames = [
     "heatmap-tracker-box",
     box.name,
@@ -19,16 +30,24 @@ export function HeatmapBox({ box }: HeatmapBoxProps) {
   ];
 
   const content =
-  box.content instanceof HTMLElement ? (
-    <span dangerouslySetInnerHTML={{ __html: box.content.outerHTML }} />
-  ) : (
-    (box.content as ReactNode)
-  );
+    box.content instanceof HTMLElement ? (
+      <span dangerouslySetInnerHTML={{ __html: box.content.outerHTML }} />
+    ) : (
+      (box.content as ReactNode)
+    );
 
   return (
     <div
       data-date={box.date}
-      style={{ backgroundColor: box.backgroundColor }}
+      data-index={index}
+      style={{
+        ...style,
+        backgroundColor: box.backgroundColor,
+        left: style.left + GUTTER_SIZE,
+        top: style.top + GUTTER_SIZE,
+        width: style.width - GUTTER_SIZE,
+        height: style.height - GUTTER_SIZE,
+      }}
       className={`${boxClassNames.filter(Boolean).join(" ")}`}
       // On Desktop it will show the date on hover.
       aria-label={box.date}
@@ -37,3 +56,5 @@ export function HeatmapBox({ box }: HeatmapBoxProps) {
     </div>
   );
 }
+
+export default HeatmapBox;