+++ src/components/SnowFall/SnowFall.tsx
@@ -1,0 +1,9 @@
+function SnowFall() {
+  return (
+    <div className="snow">
+      <div></div>
+    </div>
+  );
+}
+
+export default SnowFall;